Alll the miscellaneous bugfixes.
1) Changed CPU ticks/delta times from float to double
2) Upped VMU_DISP_GHOSTING_FRAMES from 10 to 25
3) Properly signal that LCD has changed when icons change
4) Typo fixed with OPCODE_STF_COUNT
5) Fixed bug with _biosWriteFlashRom not even ever writing to flash due
to using some obscure, old-ass unused size variable that was never set
6) Rearranged cpuUpdate order, putting interrupts and timer update first
7) Added polling gamepad to main CPU update loop
8) Changed gyVmUCpuReset register initial values based on docs

+++ include/evmu-core/gyro_vmu_osc.h
@@ -5,6 +5,8 @@
 extern "C" {
 #endif
 
+#include <stdint.h>
+
 #define VMU_OSC_QUARTZ_FREQ_HZ  32768
 #define VMU_OSC_RC_FREQ_HZ      879236
 
@@ -27,8 +29,8 @@ typedef enum VMU_OSC_MODE {
 } VMU_OSC_MODE;
 
 
-unsigned gyVmuOscCyclesPerSec(struct VMUDevice* dev);
-float gyVmuOscSecPerCycle(struct VMUDevice* dev);
+uint64_t gyVmuOscCyclesPerSec(struct VMUDevice* dev);
+double gyVmuOscSecPerCycle(struct VMUDevice* dev);
 
 //battery-life calculations
 //set/gets
